changing the rule names to match with the corresponding agda constructor names. this entails a change in naming convention, unfortunately: in agda, it's common to name functions with lower case and hyphens as in `subst-commute` or similar, and name constructors in the ML tradition with all caps as in `LCons`. so i've followed that convention.
roughly speaking, the first `n-1` letters in the all-caps prefix of a
constructor name denote the judgement it comes from, and the rest denotes
the form it applies to. if more than one rule applies to the same form,
then they're indexed at the end of the constructor name with numerals
starting at 1. so, for instance, `STAp1` is the first stepping rule for
applications because `ST` denotes the stepping judgement, `Ap` the
application form, and the numeral `1` that there are at least two rules in
that judgement that apply to applications.